Remove unused files and moved searchbar folder to its own folder
